Split ICD-10 known differences into modules
