+++ coding_systems/icd10/known_diffs/__init__.py
@@ -1,0 +1,28 @@
+from .combined2016_vs_who2019_term_differences import (
+    get_2016_2019_description_difference,
+)
+from .who2016_vs_nhs2016_code_overrides import expand_who_2016_place_of_occurrence
+from .who2016_vs_nhs2016_codes_only_in_nhs import (
+    is_2016_scraped_only,
+    should_include_2016_scraped_only,
+)
+from .who2016_vs_nhs2016_codes_only_in_who import (
+    is_2016_claml_only,
+    should_include_2016_claml_only,
+)
+from .who2016_vs_nhs2016_term_differences import (
+    is_2016_description_difference,
+    should_use_scraped_for_2016,
+)
+
+
+__all__ = [
+    "get_2016_2019_description_difference",
+    "expand_who_2016_place_of_occurrence",
+    "is_2016_scraped_only",
+    "should_include_2016_scraped_only",
+    "is_2016_claml_only",
+    "should_include_2016_claml_only",
+    "is_2016_description_difference",
+    "should_use_scraped_for_2016",
+]
