+++ coding_systems/icd10/known_diffs/combined2016_vs_who2019_term_differences.py
@@ -1,231 +1,4 @@
-# A record of known differences between:
-# - 2016 CLAML from WHO website and the 2016 CLAML generated by scraping the NHS classbrowser
-# - the NHS specific logic for things like place of occurrence modifiers
-# - the 2016 scraped and 2019 CLAML from WHO website
-
-import re
-from dataclasses import dataclass
-
-from coding_systems.icd10.claml_parser import ICD10Code, ModifierDigit
-
-
-@dataclass(frozen=True, slots=True)
-class CodeDifference:
-    include_in_release: bool
-
-
-@dataclass(frozen=True, slots=True)
-class TermDifference:
-    claml: str
-    scraped: str
-    use: str
-
-    def __post_init__(self) -> None:
-        if self.use not in ("claml", "scraped"):
-            raise ValueError(f"use must be 'claml' or 'scraped', got {self.use!r}")
-
-
-@dataclass(frozen=True, slots=True)
-class KnownDifferences:
-    claml_only: dict[str, CodeDifference]
-    scraped_only: dict[str, CodeDifference]
-    term_differences: dict[str, TermDifference]
-
-
-@dataclass(frozen=True, slots=True)
-class ReleaseTermDifference:
-    combined_2016: str
-    who_2019: str
-    clinically_equivalent: bool
-
-
-WHO_2016_EXPECTED_OVERRIDES = frozenset(
-    # Taken from https://classbrowser.nhs.uk/ICD-10-5TH-Edition/NCCS.pdf#page=218
-    #
-    # A fourth character must be assigned with codes from categories W00-Y34 to identify
-    # where the injury, poisoning or adverse effect took place. The fourth characters can
-    # be found in the ‘Place of occurrence code’ section at the beginning of the chapter.
-    # The exceptions are codes in categories Y06.- Neglect and abandonment and
-    # Y07.- Other maltreatment.
-    #
-    # However it then goes on to say in
-    #  https://classbrowser.nhs.uk/ICD-10-5TH-Edition/NCCS.pdf#page=221
-    #
-    # The fourth character codes printed at categories W26, X34, and X59 in the
-    # ICD-10 Tabular List must not be used and must be crossed through in the ICD-10 5th
-    # Edition books.
-    # The ‘Place of occurrence codes’ must be used for fourth character code assignment with
-    # categories W26.-, X34.- and X59.-.
-    # The WHO Update Revision Committee previously agreed that ‘Place of occurrence codes’
-    # should be separated from the three character Chapter XX code. Following this agreement
-    # WHO introduced new fourth character subcategory codes at W26.-, X34.- and X59.- . This
-    # re-designation and re-use of the fourth character by the WHO is incompatible with current
-    # UK practice when using categories W26.-, X34.- and X59.-, therefore the implementation
-    # of these fourth character subcategories has been deferred.
-    #
-    # This means when applying the place of occurrence modifiers we would expect the following
-    # actual 4 character children to be overridden by the place of occurrence modifiers. We
-    # want to confirm that these are the only ones, and that no other unexpected codes are
-    # generated by the place of occurrence modifiers.
-    {"W260", "W268", "W269", "X340", "X341", "X348", "X349", "X590", "X599"}
-)
-
-
-def expand_who_2016_place_of_occurrence(
-    records: dict[str, ICD10Code],
-    place_modifiers: list[ModifierDigit],
-) -> dict[str, ICD10Code]:
-    """
-    Expand the given records with place of occurrence modifiers for the 2016 WHO release,
-    and check that the expected overrides are present and no unexpected codes are generated.
-    """
-
-    place_records = {}
-    for base_code, base_record in records.items():
-        # We only apply to 3 character codes in the specified range
-        if not re.match(r"^[A-Z][0-9]{2}$", base_code):
-            continue
-
-        # We only apply to codes in the range W00-Y34
-        if not ("W00" <= base_code <= "Y34"):
-            continue
-
-        # Apart from the exceptions Y06 and Y07
-        if base_code in {"Y06", "Y07"}:
-            continue
-
-        for modifier in place_modifiers:
-            new_code = base_code + modifier.digit_code
-            place_records[new_code] = ICD10Code(
-                code=new_code,
-                parent=base_code,
-                description=base_record.description,
-                term_modifier=modifier.description,
-                modifier_position=4,
-            )
-
-    overlap = set(place_records) & set(records)
-    unexpected_overlap = overlap - WHO_2016_EXPECTED_OVERRIDES
-    assert not unexpected_overlap, (
-        "Place-of-occurrence generated codes overlap with unexpected existing codes: "
-        f"{sorted(unexpected_overlap)}"
-    )
-    unexpected_missing = WHO_2016_EXPECTED_OVERRIDES - set(place_records)
-    assert not unexpected_missing, (
-        "Place-of-occurrence expected codes missing from generated codes: "
-        f"{sorted(unexpected_missing)}"
-    )
-    return place_records
-
-
-# Known differences between the 2016 WHO ICD10 claml and the scraped ICD10
-# data from the NHS class browser
-CLAML_VS_SCRAPED_DIFFERENCES = KnownDifferences(
-    claml_only={
-        # The CLAML says you can apply the 5th char modifier to M14.*, but
-        # doesn't say you can in any online version. It also doesn't appear
-        # in the backend data. But no harm in including, and it is technically
-        # in the XML, so we keep it. The 5th char modifiers are 0-9 and apply
-        # to M140-M146 and M148-M149, but not M147 which doesn't exist
-        **{
-            f"M{i}": CodeDifference(include_in_release=True)
-            for i in list(range(1400, 1470)) + list(range(1480, 1490))
-        },
-        # W00-Y34 (except Y06 and Y07) are all 3 char codes in the CLAML,
-        # which the NHS allows a 4th character "place of occurrence" modifier.
-        # The online browser metnions this, but not in a structure way so
-        # the scraper doesn't pick it up. We therefore expect all codes
-        # in that range to be in the claml but not the scraped data.
-        # W000-W999
-        **{
-            f"W{number:03d}": CodeDifference(include_in_release=True)
-            for number in range(0, 1000)
-        },
-        # X000-X999
-        **{
-            f"X{number:03d}": CodeDifference(include_in_release=True)
-            for number in range(0, 1000)
-        },
-        # Y000-Y349 (except Y06 and Y07)
-        **{
-            f"Y{number:03d}": CodeDifference(include_in_release=True)
-            for number in list(range(0, 600)) + list(range(800, 3500))
-        },
-    },
-    scraped_only={
-        # Currently no codes that only appear in the scraped data but keeping
-        # the property here in case we want to add some in future
-    },
-    term_differences={
-        # The expected differences in the terms between the CLAML and scraped
-        # data, and which one we want to use in the release. The scraped term
-        # is usually the one we prefer as we're trying to match the NHS classbrowser
-        # as closely as possible, but in some cases the CLAML term is better.
-        "J10": TermDifference(
-            claml="influenza due to identified seasonal influenza virus",
-            scraped="Influenza due to identified seasonal influenza virus",
-            use="scraped",
-        ),
-        "P710": TermDifference(
-            claml="Cow milk hypocalcaemia in newborn",
-            scraped="Cow's milk hypocalcaemia in newborn",
-            use="scraped",
-        ),
-        "U06": TermDifference(
-            claml="Zika virus disease",
-            scraped="Emergency use of U06",
-            use="claml",
-            # U06 isn't a code you can actually use (as it has 4 char children) so
-            # the NHS haven't updated it. So the claml version is more accurate
-            # now that U609 is Zika virus instead of emergency use
-        ),
-        "U069": TermDifference(
-            claml="Zika virus disease, unspecified",
-            scraped="Zika virus disease",
-            use="scraped",
-        ),
-        "U070": TermDifference(
-            claml="Emergency use of U07.0",
-            scraped="Vaping-related disorder",
-            use="scraped",
-        ),
-        "U071": TermDifference(
-            claml="Emergency use of U07.1",
-            scraped="COVID-19, virus identified",
-            use="scraped",
-        ),
-        "U072": TermDifference(
-            claml="Emergency use of U07.2",
-            scraped="COVID-19, virus not identified",
-            use="scraped",
-        ),
-        "U073": TermDifference(
-            claml="Emergency use of U07.3",
-            scraped="Personal history of COVID-19",
-            use="scraped",
-        ),
-        "U074": TermDifference(
-            claml="Emergency use of U07.4",
-            scraped="Post COVID-19 condition",
-            use="scraped",
-        ),
-        "U075": TermDifference(
-            claml="Emergency use of U07.5",
-            scraped="Multisystem inflammatory syndrome associated with COVID-19",
-            use="scraped",
-        ),
-        "U076": TermDifference(
-            claml="Emergency use of U07.6",
-            scraped="Need for immunization against COVID-19",
-            use="scraped",
-        ),
-        "U077": TermDifference(
-            claml="Emergency use of U07.7",
-            scraped="COVID-19 vaccines causing adverse effects in therapeutic use",
-            use="scraped",
-        ),
-    },
-)
+from .difference_classes import ReleaseTermDifference
 
 
 # Known differences between the combined 2016 data (base WHO claml + scraped data)
@@ -819,48 +592,6 @@ COMBINED_2016_VS_2019_DIFFERENCES = {
         clinically_equivalent=True,
     ),
 }
-
-
-def is_2016_claml_only(code: str) -> bool:
-    """Check if a code is expected to be only in the 2016 CLAML data."""
-    return code in CLAML_VS_SCRAPED_DIFFERENCES.claml_only
-
-
-def is_2016_scraped_only(code: str) -> bool:
-    """Check if a code is expected to be only in the 2016 scraped data."""
-    return code in CLAML_VS_SCRAPED_DIFFERENCES.scraped_only
-
-
-def is_2016_description_difference(
-    code: str,
-    claml_description: str,
-    scraped_description: str,
-) -> bool:
-    """Check if a code is expected to have a different description in the 2016 CLAML vs scraped data."""
-    known = CLAML_VS_SCRAPED_DIFFERENCES.term_differences.get(code)
-    return (
-        known is not None
-        and known.claml == claml_description
-        and known.scraped == scraped_description
-    )
-
-
-def should_include_2016_claml_only(code: str):
-    """For a ClAML-only code, check if it should be included in the release."""
-    known = CLAML_VS_SCRAPED_DIFFERENCES.claml_only.get(code)
-    return known is not None and known.include_in_release
-
-
-def should_include_2016_scraped_only(code: str):
-    """For a scraped-only code, check if it should be included in the release."""
-    known = CLAML_VS_SCRAPED_DIFFERENCES.scraped_only.get(code)
-    return known is not None and known.include_in_release
-
-
-def should_use_scraped_for_2016(code: str) -> bool:
-    """For a code with a description difference, should we use the scraped description?"""
-    known = CLAML_VS_SCRAPED_DIFFERENCES.term_differences.get(code)
-    return known is not None and known.use == "scraped"
 
 
 def get_2016_2019_description_difference(
