+++ coding_systems/icd10/known_diffs/who2016_vs_nhs2016_code_overrides.py
@@ -1,0 +1,82 @@
+import re
+
+from coding_systems.icd10.claml_parser import ICD10Code, ModifierDigit
+
+
+WHO_2016_EXPECTED_OVERRIDES = frozenset(
+    # Taken from https://classbrowser.nhs.uk/ICD-10-5TH-Edition/NCCS.pdf#page=218
+    #
+    # A fourth character must be assigned with codes from categories W00-Y34 to identify
+    # where the injury, poisoning or adverse effect took place. The fourth characters can
+    # be found in the ‘Place of occurrence code’ section at the beginning of the chapter.
+    # The exceptions are codes in categories Y06.- Neglect and abandonment and
+    # Y07.- Other maltreatment.
+    #
+    # However it then goes on to say in
+    #  https://classbrowser.nhs.uk/ICD-10-5TH-Edition/NCCS.pdf#page=221
+    #
+    # The fourth character codes printed at categories W26, X34, and X59 in the
+    # ICD-10 Tabular List must not be used and must be crossed through in the ICD-10 5th
+    # Edition books.
+    # The ‘Place of occurrence codes’ must be used for fourth character code assignment with
+    # categories W26.-, X34.- and X59.-.
+    # The WHO Update Revision Committee previously agreed that ‘Place of occurrence codes’
+    # should be separated from the three character Chapter XX code. Following this agreement
+    # WHO introduced new fourth character subcategory codes at W26.-, X34.- and X59.- . This
+    # re-designation and re-use of the fourth character by the WHO is incompatible with current
+    # UK practice when using categories W26.-, X34.- and X59.-, therefore the implementation
+    # of these fourth character subcategories has been deferred.
+    #
+    # This means when applying the place of occurrence modifiers we would expect the following
+    # actual 4 character children to be overridden by the place of occurrence modifiers. We
+    # want to confirm that these are the only ones, and that no other unexpected codes are
+    # generated by the place of occurrence modifiers.
+    {"W260", "W268", "W269", "X340", "X341", "X348", "X349", "X590", "X599"}
+)
+
+
+def expand_who_2016_place_of_occurrence(
+    records: dict[str, ICD10Code],
+    place_modifiers: list[ModifierDigit],
+) -> dict[str, ICD10Code]:
+    """
+    Expand the given records with place of occurrence modifiers for the 2016 WHO release,
+    and check that the expected overrides are present and no unexpected codes are generated.
+    """
+
+    place_records = {}
+    for base_code, base_record in records.items():
+        # We only apply to 3 character codes in the specified range
+        if not re.match(r"^[A-Z][0-9]{2}$", base_code):
+            continue
+
+        # We only apply to codes in the range W00-Y34
+        if not ("W00" <= base_code <= "Y34"):
+            continue
+
+        # Apart from the exceptions Y06 and Y07
+        if base_code in {"Y06", "Y07"}:
+            continue
+
+        for modifier in place_modifiers:
+            new_code = base_code + modifier.digit_code
+            place_records[new_code] = ICD10Code(
+                code=new_code,
+                parent=base_code,
+                description=base_record.description,
+                term_modifier=modifier.description,
+                modifier_position=4,
+            )
+
+    overlap = set(place_records) & set(records)
+    unexpected_overlap = overlap - WHO_2016_EXPECTED_OVERRIDES
+    assert not unexpected_overlap, (
+        "Place-of-occurrence generated codes overlap with unexpected existing codes: "
+        f"{sorted(unexpected_overlap)}"
+    )
+    unexpected_missing = WHO_2016_EXPECTED_OVERRIDES - set(place_records)
+    assert not unexpected_missing, (
+        "Place-of-occurrence expected codes missing from generated codes: "
+        f"{sorted(unexpected_missing)}"
+    )
+    return place_records
